ml_module/requirements.txt
elasticsearch==7.15.2
aiogram==3.0.0
scikit-learn==1.4.0
pandas==2.0.3
numpy==1.26.0
python-dotenv==1.0.0
elasticsearch-async==6.2.0
apscheduler==3.10.4
pytz==2024.1